requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.80
botocore==1.35.80
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
Django==5.1.2
django-crispy-forms==2.3
django-environ==0.11.2
django-storages==1.14.4
h11==0.14.0
idna==3.10
jmespath==1.0.1
lxml==5.3.0
numpy==2.1.3
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.4
selenium==4.27.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
